+++ src/main/java/com/github/lipinskipawel/maelstrom/internal/EventDeserializer.java
@@ -29,9 +29,7 @@ final class EventDeserializer extends JsonDeserializer<Event<?>> {
         final var bodyNode = tree.get("body");
         final var intoType = supportedTypes.get(bodyNode.get("type").asText());
 
-        // either way is correct
         final var object = p.getCodec().readValue(bodyNode.traverse(), intoType);
-        // final var object = (EventType) bodyNode.traverse(p.getCodec()).readValueAs(intoType);
 
         return createEvent(id, src, dst, object);
     }